port Tests

+++ AtCoderLibrary/Math/MathLib.cs
@@ -520,7 +520,7 @@ namespace AtCoder
         /// <para>制約: gcd(<paramref name="x"/>,<paramref name="m"/>)=1, 1≤<paramref name="m"/></para>
         /// <para>計算量: O(log<paramref name="m"/>)</para>
         /// </remarks>
-        public static long InvMod(long x, int m)
+        public static long InvMod(long x, long m)
         {
             Debug.Assert(1 <= m);
             var (g, res) = InternalMath.InvGCD(x, m);
